+++ src/graphql_clj/validator/rules/no_undefined_variables.clj
@@ -12,7 +12,7 @@
 
 (defnodevisitor undefined-variable :pre :argument
   [{:keys [variable-name] :as n} s]
-  (when (and variable-name (not (s/get-spec (spec/spec-for {:node-type :variable-usage :variable-name variable-name} s))))
+  (when (and variable-name (not (s/get-spec (spec/spec-for-var-usage variable-name s))))
     {:state (ve/update-errors s (undefined-variable-error n))
      :break true}))